requirements.txt
beautifulsoup4==4.12.3
playwright==1.41.2
pypdf2==3.0.1
openai-whisper==20231117
faiss-cpu==1.7.4
python-dotenv==1.0.0
fastapi==0.109.2
uvicorn==0.27.1
torch==2.2.0
transformers==4.37.2
sentence-transformers==2.5.1
google-generativeai==0.3.2
python-multipart==0.0.6
langchain==0.0.208
pinecone-client==2.2.2
bge-reranker==0.1.0